fix: 마이페이지 코드 수정

+++ src/pages/MyPage.jsx
@@ -4,10 +4,11 @@ import { useForm } from 'react-hook-form';
 import styled from 'styled-components';
 import { useQuery, useMutation, useQueryClient } from 'react-query';
 
-import useAlert from '@/stores/useAlert';
+import useAlert from '@/hooks/modal/useAlert';
 import { useAuth } from '@components/navigation/AuthContext';
 
 import { Text } from '@/components/typograph/Text';
+import { Button } from '@components/forms/Button';
 import { Alert } from '@components/forms/modal/Alert';
 
 import defaultProfilePic from '@/assets/icons/menu/User.png';
@@ -15,13 +16,16 @@ import defaultProfilePic from '@/assets/icons/menu/User.png';
 import { API } from '@/utils/api';
 
 const Container = styled.div`
-  background-color: var(--background-color);
-  color: var(--primary-text-color);
   margin: 0;
-  margin-top: 80px;
   padding-top: 100px;
+
   display: flex;
+  flex-direction: row;
   justify-content: center;
+  align-items: center;
+
+  background-color: var(--body-background);
+  color: var(--primary-text-color);
 `;
 
 const MyPageContainer = styled.div`
@@ -32,27 +36,18 @@ const MyPageContainer = styled.div`
 `;
 
 const Section = styled.div`
-  background-color: var(--container-secondary-background);
+  background-color: var(--container-primary-background);
+  border: 1px solid var(--container-border);
+
   padding: 40px;
   border-radius: 20px;
   margin-bottom: 30px;
-  box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.3);
-`;
-
-const SectionTitle = styled.h2`
-  font-size: 24px;
-  margin-bottom: 20px;
-
-  .section-title-en {
-    font-size: 14px;
-    color: var(--primary-text-color);
-    margin-left: 10px;
-  }
 `;
 
 const ProfilePicContainer = styled.div`
   display: flex;
   align-items: center;
+  margin-top: 20px;
   margin-bottom: 20px;
 `;
 
@@ -66,24 +61,29 @@ const ProfilePic = styled.img`
 const PicButtons = styled.div`
   display: flex;
   gap: 10px;
+  height: 45px;
 
-  .change-pic-btn {
-    background-color: #3b82f6;
-    color: var(--primary-text-color);
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+
+  .change-pic-btn,
+  .delete-pic-btn {
+    width: 10vw;
+    background-color: var(--primary-color);
+    color: white;
     padding: 10px 20px;
     border: none;
     border-radius: 8px;
     cursor: pointer;
+    font-size: 14px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
   }
 
   .delete-pic-btn {
-    background-color: #ff4d4d;
-    color: var(--primary-text-color);
-    padding: 10px 20px;
-    border: none;
-    border-radius: 8px;
-    cursor: pointer;
-    font-size: 16px;
+    background-color: #D32F2F;
   }
 `;
 
@@ -95,14 +95,35 @@ const Form = styled.form`
   flex-direction: column;
 `;
 
-const InputRow = styled.div`
+const InputGroup = styled.div`
+  width: 100%;
+
   display: flex;
-  gap: 38px;
-  align-items: center;
+  gap: 20px;
+
+  & > * {
+    width: 100%;
+  }
 `;
 
-const InputGroup = styled.div`
-  width: 47%;
+const WarningMessage = styled.p`
+  font-size: 14px;
+  color: #ff6b6b;
+  margin-bottom: 20px;
+`;
+
+const InputWrapper = styled.div`
+  & > span {
+    margin-left: 10px;
+    margin-bottom: 8px;
+  }
+
+  .error-message {
+    margin-left: 10px;
+    color: var(--danger-color);
+    font-size: 14px;
+    margin-top: 10px;
+  }
 
   label {
     font-size: 16px;
@@ -111,12 +132,18 @@ const InputGroup = styled.div`
   }
 
   input {
+    transition:
+      border-color 0.2s ease-out,
+      background-color 0.2s ease-out;
     width: 100%;
-    padding: 10px;
+
+    box-sizing: border-box;
+    border-radius: 14px;
+    padding: 15px;
     margin-bottom: 20px;
-    background-color: var(--container-secondary-background);
-    border: none;
-    border-radius: 10px;
+
+    background-color: var(--container-primary-background);
+    border: 1px solid var(--container-border);
     color: var(--primary-text-color);
     outline: none;
 
@@ -125,58 +152,11 @@ const InputGroup = styled.div`
       color: var(--primary-text-color);
     }
     &:focus {
-      border: 1px solid #4a90e2;
-      box-shadow: none;
+      border: 1px solid var(--primary-color); /* 파란색 테두리 */
+      box-shadow: none; /* 흰색 테두리 제거 */
+      background-color: var(--container-secondary-background);
     }
   }
-`;
-
-const InputGroupLong = styled.div`
-  width: 98%;
-  label {
-    font-size: 16px;
-    margin-bottom: 10px;
-    display: block;
-  }
-
-  input {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 20px;
-    background-color: var(--container-secondary-background);
-    border: var(--container-border);
-    border-radius: 10px;
-    color: var(--primary-text-color);
-    outline: none;
-
-    &:read-only {
-      background-color: #2c2f3e;
-      color: var(--primary-text-color);
-      outline: none;
-    }
-
-    &:focus {
-      border: 1px solid #4a90e2;
-      box-shadow: none;
-  }
-`;
-
-const EditButton = styled.button`
-  background-color: #3b82f6;
-  color: var(--primary-text-color);
-  padding: 12px 20px;
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  width: 200px;
-  margin-left: auto;
-  matgin-top: auto;
-`;
-
-const WarningMessage = styled.p`
-  font-size: 14px;
-  color: #ff6b6b;
-  margin-bottom: 20px;
 `;
 
 export default function MyPage() {
@@ -193,7 +173,6 @@ export default function MyPage() {
   const [passwordError, setPasswordError] = useState('');
   const navigate = useNavigate();
   const { openAlert, closeAlert, isOpen } = useAlert();
-  const queryClient = useQueryClient();
 
   const {
     register,
@@ -204,12 +183,12 @@ export default function MyPage() {
 
   // 로그인 여부 확인
   useEffect(() => {
-    if (!isLoggedIn) {
+    if (isLoggedIn) {
       navigate('/login');
     }
   }, [isLoggedIn, navigate]);
 
-  // Fetch user data after login
+  // Fetch user data
   const { isLoading } = useQuery(
     ['userData', user?.student_id],
     async () => {
@@ -223,23 +202,25 @@ export default function MyPage() {
     },
     {
       enabled: isLoggedIn,
-      onSuccess: (data) => setUserInfo({
-        studentNumber: data.student_id,
-        name: data.name,
-        email: data.email,
-        generation: data.generation,
-        major: data.major,
-        profilePic: data.profile_picture || defaultProfilePic,
-      }),
-      onError: (error) => {
-        console.error('Error fetching user data:', error); // 콘솔에서 오류 확인
+      onSuccess: (data) =>
+        setUserInfo({
+          studentNumber: data.student_id,
+          name: data.name,
+          email: data.email,
+          generation: data.generation,
+          major: data.major,
+          profilePic: data.profile_picture || defaultProfilePic,
+        }),
+      onError: () => {
         openAlert({
           title: '사용자 정보 불러오기 실패',
-          content: <Text>사용자 정보 불러오기에 실패했습니다. 다시 시도해주세요.</Text>,
+          content: (
+            <Text>사용자 정보 불러오기에 실패했습니다. 다시 시도해주세요.</Text>
+          ),
           onClose: closeAlert,
         });
       },
-    }
+    },
   );
 
   const imageUploadMutation = useMutation(
@@ -261,17 +242,21 @@ export default function MyPage() {
           };
 
           try {
-            const response = await API.PUT(`/users/${userInfo.student_id}`, formData, {
-              headers: {
-                Authorization: `Bearer ${token}`,
+            const response = await API.PUT(
+              `/users/${userInfo.student_id}`,
+              formData,
+              {
+                headers: {
+                  Authorization: `Bearer ${token}`,
+                },
               },
-            });
+            );
             resolve(response.data);
           } catch (error) {
             reject(error);
           }
         };
-        reader.onerror = () => reject(new Error("이미지 변환 실패"));
+        reader.onerror = () => reject(new Error('이미지 변환 실패'));
       });
     },
     {
@@ -282,11 +267,13 @@ export default function MyPage() {
       onError: () => {
         openAlert({
           title: '이미지 업로드 실패',
-          content: <Text>이미지 업로드에 실패했습니다. 다시 시도해주세요.</Text>,
+          content: (
+            <Text>이미지 업로드에 실패했습니다. 다시 시도해주세요.</Text>
+          ),
           onClose: closeAlert,
         });
       },
-    }
+    },
   );
 
   const handleImageUpload = (event) => {
@@ -318,60 +305,61 @@ export default function MyPage() {
       onError: () => {
         openAlert({
           title: '비밀번호 재설정 실패',
-          content: <Text>비밀번호 재설정에 실패했습니다. 다시 시도해주세요.</Text>,
+          content: (
+            <Text>비밀번호 재설정에 실패했습니다. 다시 시도해주세요.</Text>
+          ),
           onClose: closeAlert,
         });
       },
-    }
+    },
   );
 
-  const onSubmit = async (data) => {
+  const onSubmit = (data) => {
     setPasswordError('');
     if (data.newPassword !== data.confirmPassword) {
       setPasswordError('새로운 비밀번호가 일치하지 않습니다.');
       return;
     }
+
     // hash를 생성하여 data 객체를 수정
-    const hashData = { hash: data.newPassword };
+    const hashData = {
+      currentPassword: data.currentPassword, // 기존 비밀번호 포함
+      newPassword: data.newPassword, // 새 비밀번호
+    };
+
+    // 비밀번호 변경 요청
     passwordChangeMutation.mutate(hashData);
   };
 
-  const deleteAccountMutation = useMutation(
-    async () => {
-      const token = localStorage.getItem('accessToken');
-      return await API.DELETE(`/users/${user.student_id}`, {
-        headers: { Authorization: token },
-      });
-    },
-    {
-      onSuccess: () => {
+  const handleDeleteAccount = () => {
+    const confirmDelete = window.confirm(
+      '계정을 삭제하면 복구할 수 없습니다. 정말로 삭제하시겠습니까?',
+    );
+    if (!confirmDelete) return;
+    // 계정 삭제 요청
+    const token = localStorage.getItem('accessToken');
+    API.DELETE(`/users/${user.student_id}`, {
+      headers: {
+        Authorization: token,
+      },
+    })
+      .then(() => {
         openAlert({
-          title: '계정 삭제',
+          title: '계정 삭제 성공',
           content: <Text>계정이 성공적으로 삭제되었습니다.</Text>,
           onClose: () => {
-            closeAlert();
             logout();
             navigate('/login');
           },
         });
-      },
-      onError: () => {
+      })
+      .catch(() => {
         openAlert({
           title: '계정 삭제 실패',
           content: <Text>계정 삭제에 실패했습니다. 다시 시도해주세요.</Text>,
           onClose: closeAlert,
         });
-      },
-    }
-  );
-
-  const handleDeleteAccount = () => {
-    openAlert({
-      title: '계정 삭제 확인',
-      content: <Text>계정을 삭제하면 복구할 수 없습니다. 정말로 삭제하시겠습니까?</Text>,
-      ok_label: '확인',
-      onClose: deleteAccountMutation.mutate,
-    });
+      });
   };
 
   return (
@@ -379,9 +367,12 @@ export default function MyPage() {
       <MyPageContainer>
         {/* Account Info Section */}
         <Section>
-          <SectionTitle>
-            계정 정보 <span className="section-title-en">Account Info</span>
-          </SectionTitle>
+          <Text size="m" weight="light" color="--secondary-text-color">
+            Account Info
+          </Text>
+          <Text size="sxl" weight="bold">
+            계정 정보
+          </Text>
           <ProfilePicContainer>
             <ProfilePic src={userInfo.profilePic} alt="Profile" />
             <PicButtons>
@@ -404,40 +395,44 @@ export default function MyPage() {
             </PicButtons>
           </ProfilePicContainer>
           <Form>
-            <InputRow>
-              <InputGroup>
+            <InputGroup>
+              <InputWrapper>
                 <label htmlFor="name">이름</label>
                 <input
                   type="text"
+                  placeholder="이름"
                   id="name"
                   name="name"
                   value={userInfo.name}
                   readOnly
                 />
-              </InputGroup>
+              </InputWrapper>
 
-              <InputGroup>
+              <InputWrapper>
                 <label htmlFor="studentNumber">학번</label>
                 <input
                   type="text"
+                  placeholder="학번"
                   id="studentNumber"
                   name="studentNumber"
                   value={userInfo.studentNumber}
                   readOnly
                 />
-              </InputGroup>
-            </InputRow>
+              </InputWrapper>
+            </InputGroup>
           </Form>
         </Section>
 
         {/* Change Password Section */}
         <Section>
-          <SectionTitle>
-            비밀번호 변경{' '}
-            <span className="section-title-en">Change Password</span>
-          </SectionTitle>
+          <Text size="m" weight="light" color="--secondary-text-color">
+            Change Password
+          </Text>
+          <Text size="sxl" weight="bold">
+            비밀번호 변경
+          </Text>
           <Form onSubmit={handleSubmit(onSubmit)}>
-            <InputGroupLong>
+            <InputWrapper>
               <label htmlFor="current-password">현재 비밀번호 입력</label>
               <input
                 type="password"
@@ -458,9 +453,9 @@ export default function MyPage() {
                   {errors.currentPassword.message}
                 </WarningMessage>
               )}
-            </InputGroupLong>
-            <InputRow>
-              <InputGroup>
+            </InputWrapper>
+            <InputGroup>
+              <InputWrapper>
                 <label htmlFor="new-password">새 비밀번호 입력</label>
                 <input
                   type="password"
@@ -474,7 +469,7 @@ export default function MyPage() {
                     },
                     pattern: {
                       value:
-                        /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*]).{8,20}$/,
+                        /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*])[A-Za-z\d!@#$%^&*]{8,20}$/,
                       message:
                         '비밀번호는 숫자, 영문 대문자·소문자, 특수문자를 포함해야 합니다.',
                     },
@@ -483,9 +478,9 @@ export default function MyPage() {
                 {errors.newPassword && (
                   <WarningMessage>{errors.newPassword.message}</WarningMessage>
                 )}
-              </InputGroup>
+              </InputWrapper>
 
-              <InputGroup>
+              <InputWrapper>
                 <label htmlFor="confirm-password">새 비밀번호 확인</label>
                 <input
                   type="password"
@@ -503,27 +498,38 @@ export default function MyPage() {
                     {errors.confirmPassword.message}
                   </WarningMessage>
                 )}
-              </InputGroup>
-            </InputRow>
+              </InputWrapper>
+            </InputGroup>
 
             {passwordError && <WarningMessage>{passwordError}</WarningMessage>}
 
-            <EditButton type="submit">비밀번호 변경</EditButton>
+            <div style={{ display: 'flex', justifyContent: 'flex-end'}}>
+              <Button width="25%" height="45px">
+                비밀번호 변경
+              </Button>
+            </div>
           </Form>
         </Section>
 
         {/* Delete Account Section */}
         <Section>
-          <SectionTitle>
-            계정 삭제 <span className="section-title-en">Delete Account</span>
-          </SectionTitle>
-          <Form>
+          <Text size="m" weight="light" color="--secondary-text-color">
+            Delete Account
+          </Text>
+          <Text size="sxl" weight="bold">
+            계정 삭제
+          </Text>
+
+          <Form onClick={handleDeleteAccount}>
             <WarningMessage>
               계정을 삭제하면 복구할 수 없습니다. 신중히 선택하세요.
             </WarningMessage>
-            <EditButton type="button" onClick={handleDeleteAccount}>
-              계정 삭제
-            </EditButton>
+
+            <div style={{ display: 'flex', justifyContent: 'flex-end'}}>
+              <Button width="25%" height="45px">
+                계정 삭제
+              </Button>
+            </div>
           </Form>
         </Section>
         <Alert isOpen={isOpen} closeAlert={closeAlert} />
